libgupnp-dlna: Detect duplicates in <include>
Currently, if we include one of the XML files in the profiles directory
from another XML file, the former will get parsed twice (once while
walking the directory, and once again when being included), resulting in
the same set profiles being added to the profile list twice.

This adds some (not fool-proof, but good enough) duplicate detection, so
that we can include one XML file in another. This is particularly useful
for reusing audio profile restrictions in video profiles.