libgupnp-dlna: Move profile list to DLNADiscovererClass
Instead of keeping the profile list loaded from disk in a static global
variable, we now make this a static member on the GUPnPDLNADiscoverer
class.

Corresponding to this, we move _get_profile() (and future API of a
similar nature) from the library-wide namespace into the class'
namespace.

Based on a patch from:
Parthasarathi Susarla <partha.susarla@collabora.co.uk>